refactor: 팩토리 메소드 패턴 적용해서 getConnection 분리
