+++ NUserDao.java
@@ -1,0 +1,28 @@
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class NUserDao extends UserDao {
+    public static void main(String[] args) throws SQLException, ClassNotFoundException {
+        UserDao dao = new NUserDao();
+
+        User user = new User();
+        user.setId("hannkim");
+        user.setName("김한나");
+        user.setPassword("genius");
+
+        dao.add(user);
+
+        System.out.println(user.getId() + " 등록 성공");
+
+        User user2 = dao.get(user.getId());
+        System.out.println(user2.getName());
+
+    }
+
+    public Connection getConnection() throws ClassNotFoundException, SQLException {
+        Class.forName("com.mysql.cj.jdbc.Driver");
+        Connection c = DriverManager.getConnection("jdbc:mysql://localhost/springbook", "spring", "book1234");
+        return c;
+    }
+}
