+++ UserDao.java
@@ -1,64 +1,42 @@
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+public abstract class UserDao {
+    public void add(User user) throws ClassNotFoundException, SQLException {
+        Connection c = getConnection();
 
-public class UserDao {
-	public void add(User user) throws ClassNotFoundException, SQLException {
-		Connection c = getConnection();
+        PreparedStatement ps = c.prepareStatement("insert into users(id, name, password) values(?,?,?)");
+        ps.setString(1, user.getId());
+        ps.setString(2, user.getName());
+        ps.setString(3, user.getPassword());
 
-		PreparedStatement ps = c.prepareStatement("insert into users(id, name, password) values(?,?,?)");
-		ps.setString(1, user.getId());
-		ps.setString(2, user.getName());
-		ps.setString(3, user.getPassword());
+        ps.executeUpdate();
 
-		ps.executeUpdate();
+        ps.close();
+        c.close();
+    }
 
-		ps.close();
-		c.close();
-	}
+    public User get(String id) throws ClassNotFoundException, SQLException {
+        Connection c = getConnection();
+        PreparedStatement ps = c.prepareStatement("select * from users where id = ?");
+        ps.setString(1, id);
 
-	public User get(String id) throws ClassNotFoundException, SQLException {
-		Connection c = getConnection();
-		PreparedStatement ps = c.prepareStatement("select * from users where id = ?");
-		ps.setString(1, id);
+        ResultSet rs = ps.executeQuery();
+        rs.next();
+        User user = new User();
+        user.setId(rs.getString("id"));
+        user.setName(rs.getString("name"));
+        user.setPassword(rs.getString("password"));
 
-		ResultSet rs = ps.executeQuery();
-		rs.next();
-		User user = new User();
-		user.setId(rs.getString("id"));
-		user.setName(rs.getString("name"));
-		user.setPassword(rs.getString("password"));
+        rs.close();
+        ps.close();
+        c.close();
 
-		rs.close();
-		ps.close();
-		c.close();
+        return user;
+    }
 
-		return user;
-	}
+    public abstract Connection getConnection() throws ClassNotFoundException, SQLException;
 
-	private Connection getConnection() throws ClassNotFoundException, SQLException {
-		Class.forName("com.mysql.cj.jdbc.Driver");
-		Connection c = DriverManager.getConnection("jdbc:mysql://localhost/springbook", "spring", "book1234");
-		return c;
-	}
-
-	public static void main(String[] args) throws SQLException, ClassNotFoundException {
-		UserDao dao = new UserDao();
-
-		User user = new User();
-		user.setId("hannkim");
-		user.setName("김한나");
-		user.setPassword("genius");
-
-		dao.add(user);
-
-		System.out.println(user.getId() + " 등록 성공");
-
-		User user2 = dao.get(user.getId());
-		System.out.println(user2.getName());
-
-	}
 }
